+++ Queue-online-judge/src/main/java/cn/queue/online_judge/pojo/PageBean.java
@@ -13,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PageBean {
-    private Long total;//总记录数
+    private Integer total;//总记录数
     private List rows;//数据列表
 }
