+++ Queue-online-judge/src/main/java/cn/queue/online_judge/service/ModuleService.java
@@ -5,7 +5,7 @@ import cn.queue.online_judge.pojo.Module;
 import java.util.List;
 
 public interface ModuleService {
-    Module getById(Integer id);
+    Module getById(Long id);
 
 
     void create(Module module);
